+++ src/pages/diaryDetails/DiaryDetail.jsx
@@ -1,10 +1,11 @@
 import React, { useState } from "react";
 import { dateToDotString } from "../../utils/api/dateConverter";
-import Slider from "react-slick";
 import { useCalendarStore } from "../../stores/CalendarStore";
 import useFetchDiary from "../../hooks/diary/queries/useFetchDiary";
 import { useInput } from "../../hooks/inputs/useInput";
 import TextareaAutosize from "react-textarea-autosize";
+import DiaryKeywordImagesSlider from "../../components/diaryDetail/DiaryKeywordImagesSlider";
+import DiaryControlButton from "./DiaryControlButton";
 
 export const DIARY_DETAIL_PAGE_PATH = "/diary/detail";
 
@@ -19,7 +20,7 @@ const DiaryDetail = () => {
   const [controlState, setControlState] = useState({
     isEditActive: !isDiaryExist,
     controlButtonActive: true,
-    controlButtonText: !isDiaryExist ? "작성하기" : "수정하기",
+    controlButtonText: `${!isDiaryExist ? "작성하기" : "수정하기"}`,
     controlButtonMessage: "",
   });
 
@@ -46,53 +47,6 @@ const DiaryDetail = () => {
   );
 };
 
-const settings = {
-  infinite: false,
-  arrows: false,
-  dots: true,
-  swipe: true,
-};
-
-const DiaryKeywordImagesSlider = () => {
-  let { isDiaryExist, diary } = useFetchDiary();
-
-  if (!isDiaryExist) {
-    return null;
-  }
-
-  const keywords = diary.data.result[0].keywords;
-
-  if (!keywords.some(({ imgUrl }) => imgUrl)) {
-    return <NoDiaryImage />;
-  }
-
-  return (
-    <Slider {...settings} className={"w-full rounded-xl overflow-clip"}>
-      {keywords.map(({ keywordId, imgUrl, keyword }, index) => (
-        <img
-          key={keywordId}
-          src={imgUrl}
-          alt={keyword}
-          className={`w-full bg-white aspect-square`}
-        />
-      ))}
-    </Slider>
-  );
-};
-
-const NoDiaryImage = () => {
-  return (
-    <div className={"mx-auto font-bold h-auto"}>
-      <span className={"text-2xl text-end tablet:text-4xl text-nowrap"}>
-        그려진 그림이 없어요.. <span className={"ml-4"}>😢</span>
-      </span>
-      <p className={"text-xl text-end mt-6 tablet:text-3xl cursor-pointer"}>
-        그림 그리러 가기 >
-      </p>
-    </div>
-  );
-};
-
 const DiaryDate = () => {
   const HorizontalLine = ({ className }) => {
     return <div className={`${className} border-2 border-[#5B5B5B]`}></div>;
@@ -108,61 +62,10 @@ const DiaryDate = () => {
   );
 };
 
-/**
- * @param content
- * @param className
- * @param controlState {ControlState}
- * @param setControlState
- * @param props
- * @return {Element}
- * @constructor
- */
-const DiaryControlButton = ({
-  content,
-  className,
-  controlState,
-  setControlState,
-  ...props
-}) => {
-  const handleClick = (e) => {
-    if (content === "") {
-      setControlState({
-        ...controlState,
-        message: "일기를 입력해주세요.",
-      });
-      return;
-    }
-
-    setControlState({
-      ...controlState,
-      isEditActive: false,
-      controlButtonActive: false,
-      controlButtonMessage: "",
-    });
-  };
-
-  return (
-    <>
-      <div className={"text-2xl text-red-600 text-center"}>
-        {controlState.controlButtonMessage}
-      </div>
-      <button
-        disabled={!controlState.controlButtonActive}
-        onClick={handleClick}
-        className={`border-2 bg-secondary-400 rounded-lg text-2xl py-3 ${className}`}
-        {...props}
-      >
-        {controlState.controlButtonText}
-      </button>
-    </>
-  );
-};
-
 export default DiaryDetail;
 
 /**
  * @typedef {{
- *   content: string,
  *   isEditActive: boolean,
  *   controlButtonActive: boolean,
  *   controlButtonText: string,
